Refactored with blueprints for large flask app structure.
