--- README.md
@@ -1,5 +1,0 @@
-# FlaskWeb1
-Website That Works and Does What it Should
-This is a website I was ask to make in order to understand stuff in Python.
-update.
-1
