--- create_tables.py
@@ -1,3 +1,0 @@
-from app import db
-
-db.create_all()
