+++ project/articles/templates/articles/dashboard.html
@@ -2,7 +2,7 @@
 
 {% block body %}
 <h1>Dashboard <small> Welcome {{session.username}}</small></h1>
-<a class="btn btn-success" href="/add_article"> Add Article</a>
+<a class="btn btn-success" href="/articles/add_article"> Add Article</a>
 <hr>
 <table class="table table-striped">
     <tr>
@@ -22,7 +22,7 @@
         <td>{{article.create_date}}</td>
         <td><a href="edit_article/{{article.id}}" class="btn btn-default pull-right">Edit</a></td>
         <td>
-            <form action="{{url_for('delete_article', a_id=article.id)}}" method="post">
+            <form action="{{url_for('articles.delete_article', a_id=article.id)}}" method="post">
                 <input type="hidden" name="_method" value="DELETE">
                 <input type="submit" value="Delete" class="btn btn-danger">
             </form>
@@ -36,7 +36,7 @@
         <td>{{article.create_date}}</td>
         <td><a href="edit_article/{{article.id}}" class="btn btn-default pull-right">Edit</a></td>
         <td>
-            <form action="{{url_for('delete_article', a_id=article.id)}}" method="post">
+            <form action="{{url_for('articles.delete_article', a_id=article.id)}}" method="post">
                 <input type="hidden" name="_method" value="DELETE">
                 <input type="submit" value="Delete" class="btn btn-danger">
             </form>
