+++ project/templates/includes/_navbar.html
@@ -16,9 +16,9 @@
                 <li><a href="/about">About</a></li>
                 {% if session.logged_in %}
                     {% if session.admin==1 %}
-                        <li><a href="/list_db">List DB</a></li>
+                        <li><a href="/users/list_db">List DB</a></li>
                     {% else %}
-                        <li><a href="/u_data/{{session['username']}}">User Data</a></li>
+                        <li><a href="/users/u_data/{{session['username']}}">User Data</a></li>
                     {% endif %}
                 {% endif %}
                 <li><a href="/articles">Articles</a></li>
@@ -27,11 +27,11 @@
 
             <ul class="nav navbar-nav navbar-right">
                 {% if session.logged_in %}
-                    <li><a href="/dashboard">Dashboard</a></li>
-                    <li><a href="/logout">Logout</a></li>
+                    <li><a href="/articles/dashboard">Dashboard</a></li>
+                    <li><a href="/users/logout">Logout</a></li>
                 {% else %}
-                    <li><a href="/register">Register</a></li>
-                    <li><a href="/login">Login</a></li>
+                    <li><a href="/users/register">Register</a></li>
+                    <li><a href="/users/login">Login</a></li>
                 {% endif %}
 
             </ul>
