+++ project/users/templates/users/list_db.html
@@ -20,7 +20,7 @@
         <td>{{user.email}}</td>
         <td><a href="edit_user/{{user.username}}" class="btn btn-default pull-right">Edit</a></td>
         <td>
-          <form action="{{url_for('delete_user', u_id=user.id)}}" method="post">
+          <form action="{{url_for('users.delete_user', u_id=user.id)}}" method="post">
             <input type="hidden" name="_method" value="DELETE">
             <input type="submit" value="Delete" class="btn btn-danger">
           </form>
