+++ project/users/templates/users/register.html
